setPowerdB to common
