+++ src/src/common.c
@@ -1,5 +1,6 @@
 #include "common.h"
 #include "serial.h"
+#include "targets.h"
 #include <string.h>
 
 uint8_t rxPacket[16];
@@ -22,4 +23,29 @@ void zeroRxPacket(void)
 void zeroTxPacket(void)
 {
     memset(txPacket, 0, sizeof(txPacket));
+}
+
+void setPowerdB(uint16_t currPowerdB)
+{
+  switch (currPowerdB)
+  {
+  case 0:
+    setPowermW(0);
+    break;
+  case 14:
+    setPowermW(25);
+    break;
+  case 20:
+    setPowermW(100);
+    break;
+  case 23:
+    setPowermW(200);
+    break;
+  case 26:
+    setPowermW(400);
+    break;
+  default:
+    return; // power value not recognised and no change
+    break;
+  }
 }
