+++ src/src/targets.h
@@ -24,6 +24,5 @@ void taget_setup(void);
 void taget_loop(void);
 
 void setPowermW(uint16_t power);
-void setPowerdB(uint16_t currPowerdB);
 void rfPowerAmpPinSetup(void);
 uint32_t vpd_value_get(void);
